+++ TeduCoreApp.WebApi/Controllers/ProductImageController.cs
@@ -2,11 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TeduCoreApp.Application.Interfaces;
 using TeduCoreApp.Data.ViewModels.Product;
+using TeduCoreApp.Utilities.Constants;
+using TeduCoreApp.WebApi.Authorization;
 using TeduCoreApp.WebApi.Extensions;
 
 namespace TeduCoreApp.WebApi.Controllers
@@ -16,10 +19,12 @@ namespace TeduCoreApp.WebApi.Controllers
     {
         private IProductImageService _productImageService;
         private IHostingEnvironment _env;
-        public ProductImageController(IProductImageService productImageService,IHostingEnvironment env)
+        private readonly IAuthorizationService _authorizationService;
+        public ProductImageController(IProductImageService productImageService,IHostingEnvironment env, IAuthorizationService authorizationService)
         {
             _productImageService = productImageService;
             _env = env;
+            _authorizationService = authorizationService;
         }
         
         [HttpGet]
@@ -38,8 +43,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpPost]
         [Route("add")]
-        public IActionResult Add([FromBody] ProductImageViewModel productImageVm)
+        public async Task<IActionResult> Add([FromBody] ProductImageViewModel productImageVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Create);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 _productImageService.Add(productImageVm);
@@ -50,8 +60,13 @@ namespace TeduCoreApp.WebApi.Controllers
         }
         [HttpPut]
         [Route("update")]
-        public IActionResult Update([FromBody] ProductImageViewModel productImageVm)
+        public async Task<IActionResult> Update([FromBody] ProductImageViewModel productImageVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Update);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 _productImageService.Update(productImageVm);
@@ -63,8 +78,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpDelete] 
         [Route("delete")]
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Delete);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             ProductImageViewModel productImageVm = _productImageService.GetById(id);
             _productImageService.Delete(id);
             _productImageService.SaveChanges();
