+++ TeduCoreApp.WebApi/Controllers/ProductQuantityController.cs
@@ -2,10 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TeduCoreApp.Application.Interfaces;
 using TeduCoreApp.Data.ViewModels.Product;
+using TeduCoreApp.Utilities.Constants;
+using TeduCoreApp.WebApi.Authorization;
 
 namespace TeduCoreApp.WebApi.Controllers
 {
@@ -13,10 +16,11 @@ namespace TeduCoreApp.WebApi.Controllers
     public class ProductQuantityController : ApiController
     {
         IProductQuantityService _productQuantityService;
-
-        public ProductQuantityController(IProductQuantityService productQuantityService)
+        private readonly IAuthorizationService _authorizationService;
+        public ProductQuantityController(IProductQuantityService productQuantityService, IAuthorizationService authorizationService)
         {
             _productQuantityService = productQuantityService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -42,8 +46,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpPost]
         [Route("add")]
-        public IActionResult Add([FromBody] ProductQuantityViewModel productQuantityVm)
+        public async Task<IActionResult> Add([FromBody] ProductQuantityViewModel productQuantityVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Create);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 if (_productQuantityService.CheckExist(productQuantityVm.ProductId, productQuantityVm.SizeId, productQuantityVm.ColorId))
@@ -59,8 +68,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpPut]
         [Route("update")]
-        public IActionResult Update([FromBody] ProductQuantityViewModel productQuantityVm)
+        public async Task<IActionResult> Update([FromBody] ProductQuantityViewModel productQuantityVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Update);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 _productQuantityService.Update(productQuantityVm);
@@ -72,8 +86,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpDelete]
         [Route("delete")]
-        public IActionResult Delete(int productId, int sizeId, int colorId)
+        public async Task<IActionResult> Delete(int productId, int sizeId, int colorId)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "PRODUCT", Operations.Delete);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             _productQuantityService.Delete(productId,sizeId,colorId);
             _productQuantityService.SaveChanges();
             return new OkObjectResult(productId);
