+++ TeduCoreApp.WebApi/Controllers/SendMailController.cs
@@ -2,11 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TeduCoreApp.Application.Interfaces;
 using TeduCoreApp.Data.ViewModels.Subcrible;
+using TeduCoreApp.Utilities.Constants;
 using TeduCoreApp.Utilities.Dtos;
+using TeduCoreApp.WebApi.Authorization;
 using TeduCoreApp.WebApi.ViewModel;
 
 namespace TeduCoreApp.WebApi.Controllers
@@ -16,16 +19,23 @@ namespace TeduCoreApp.WebApi.Controllers
     {
         private ISendMailService _senMailService;
         private ISubcribleService _subcribleService;
-        public SendMailController(ISendMailService senMailService, ISubcribleService subcribleService)
+        private readonly IAuthorizationService _authorizationService;
+        public SendMailController(ISendMailService senMailService, ISubcribleService subcribleService, IAuthorizationService authorizationService)
         {
             _senMailService = senMailService;
             _subcribleService = subcribleService;
+            _authorizationService = authorizationService;
         }
 
         [HttpPost]
         [Route("subcrible")]
         public async Task<IActionResult> SendMuliMail([FromBody] SendEmailViewModel sendEmailVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "USER", Operations.Read);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {              
                 try
@@ -45,8 +55,13 @@ namespace TeduCoreApp.WebApi.Controllers
         }
         [HttpGet]
         [Route("getall")]
-        public IActionResult GetSubcrible(int page,int pageSize)
+        public async Task<IActionResult> GetSubcrible(int page,int pageSize)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "USER", Operations.Read);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             List<SubcribleViewModel> listSubs = _subcribleService.GetPaging(page, pageSize, out int totalRows);
             return new OkObjectResult(new ApiResultPaging<SubcribleViewModel>()
             {
@@ -59,8 +74,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpDelete]
         [Route("delete")]
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "USER", Operations.Delete);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             _subcribleService.Delete(id);
             _subcribleService.SaveChanges();
             return new OkObjectResult(id);
