+++ TeduCoreApp.WebApi/Controllers/SystemConfigController.cs
@@ -2,11 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TeduCoreApp.Application.Interfaces;
 using TeduCoreApp.Data.Entities;
 using TeduCoreApp.Data.ViewModels.SystemConfig;
+using TeduCoreApp.Utilities.Constants;
+using TeduCoreApp.WebApi.Authorization;
 using TeduCoreApp.WebApi.Extensions;
 
 namespace TeduCoreApp.WebApi.Controllers
@@ -15,15 +18,17 @@ namespace TeduCoreApp.WebApi.Controllers
     public class SystemConfigController : ApiController
     {
         private ISystemConfigService _systemConfigService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public SystemConfigController(ISystemConfigService systemConfigService)
+        public SystemConfigController(ISystemConfigService systemConfigService, IAuthorizationService authorizationService)
         {
             _systemConfigService = systemConfigService;
+            _authorizationService = authorizationService;
         }
         [Route("getall")]
         [HttpGet]
         public IActionResult Get()
-        {
+        {           
             return new OkObjectResult(_systemConfigService.GetAll());
         }
 
@@ -36,8 +41,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [Route("add")]
         [HttpPost]
-        public IActionResult Add([FromBody] SystemConfigViewModel systemConfigVm)
+        public async Task<IActionResult> Add([FromBody] SystemConfigViewModel systemConfigVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "SYSTEMCONFIG", Operations.Create);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 _systemConfigService.Add(systemConfigVm);
@@ -49,8 +59,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [Route("update")]
         [HttpPut]
-        public IActionResult Update([FromBody] SystemConfigViewModel systemConfigVm)
+        public async Task<IActionResult> Update([FromBody] SystemConfigViewModel systemConfigVm)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "SYSTEMCONFIG", Operations.Update);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             if (ModelState.IsValid)
             {
                 SystemConfig systemConfigDb = _systemConfigService.DetailDb(systemConfigVm.Id);
@@ -64,8 +79,13 @@ namespace TeduCoreApp.WebApi.Controllers
 
         [HttpDelete]
         [Route("delete")]
-        public IActionResult Delete(string id)
+        public async Task<IActionResult> Delete(string id)
         {
+            var hasPermission = await _authorizationService.AuthorizeAsync(User, "SYSTEMCONFIG", Operations.Delete);
+            if (hasPermission.Succeeded == false)
+            {
+                return new BadRequestObjectResult(CommonConstants.Forbidden);
+            }
             _systemConfigService.Delete(id);
             _systemConfigService.SaveChanges();
             return new OkObjectResult(id);
